Product variants -Moved test from variants to configurable variants and other minor adjustments
